Both i2c busses are now working for DACs.
Remaining problem: actually writing to the DACs (or perhaps reading the values from Serial) causes the sample rate to drop from 1kHz to 1Hz. I'm pretty certain the 1Hz thing is a timeout, which means that Serial.parseInt() is waiting to finish reading an incoming integer. I can't say why that's happening. The 8 DAC values sent by the Easel seem to be formatted correctly.