All functions are now proven!
ADC -> Easel is good
DAC -> Palette is good
DAC on both i2c busses is good
DAC -> Servos is good!

Weird:
Servo is not rotating fully. This is, I think, in the firmware.
Sample rate is still pretty low. I'll have to see what it takes to bundle up more samples at once. [group 0] with a [count 2] to trigger it will, I think, get a double list of the DAC values on the Easel. Palette will need a parallel process.