More input and output channel corrections
Reduced the white LED inputs to have a range of 0V:3.3V to get a greater range of brightness across the range of values

Powered the input op amps from 3v3 so we can't blow up. Changed the comparison voltage for the LEDs to 0V:3.3V instead of up to 5 so brightnesses are consistent between input and output channels.