Fixed the lack of outgoing data to the Palette
It looks like I'd stopped feeding the output to [pd formatPacket], probably as part of an earlier bug hunt.

Currently, you have to click [open 3] to start communication. The switch on the panel is disconnected, as it tries to turn itself on, leading to a stack overflow due to infinite loop.