Hunting for bugs between Linux and Mac
Palette works fine on Mac, though with a 40Hz sample rate.

Palette works in Linux with a 340Hz sample rate but suddenly stops after a random number of seconds. It does *not* go down to a sample rate of 0.25 as it was doing before. The microcontroller just completely stops at a certain point in the process.